+++ packages/mangrove.js/CHANGELOG.md
@@ -1,5 +1,7 @@
 # next version
 
+# 0.9.0 (August 2022)
+
 - several bugfixes
 - [strats] Providing an 'activate' function to enable strat admin to set required allowances in a single transaction
 - Providing access to `activate` in the API (`OfferLogic` class)
@@ -7,6 +9,7 @@
 - [major update] adding `enableRouting` function on `liquidityProvider` class to approve router to route liquidity on current market.
 - `cancelOffer/Bid/Ask` is now called `retractOffer/Bid/Ask` to match Mangrove's naming scheme
 - Remove deprecated and defunct `printOrderBook.ts` script which has been superseded by the `mgv print` commmand
+- bump commonlib.js to 0.0.4 as commonlib.js@0.0.2 was broken
 
 # 0.8.1 (June 2022)